explicitly test for systemd presence
Journalctl won't work properly on systems not being initialized
using systemd init system. It can be tricked to work by starting
systemd logger service explicitly but the test does not assume that.

Containers with systemd entrypoint needs to be running in privileged
mode (as of today) which implies that most of the GitLab CI runners
won't be able to run systemd-enabled containers.

This commit is therefore properly detecting systemd rather than
just assuming that it is not available in the CI pipelines only.