[#5] If YottaDB was installed with UTF-8, always build both M and UTF-8 mode versions
This commit modifies CMakeLists to detect if YottaDB is installed with support for UTF-8. If UTF-8 mode support is installed, then the plugin is also built with UTF-8 support without having to run the installer twice. This is done by adding a file called ydbposixso.cmake which now handles the building of _ydbposix.so files. This file is based on sr_unix/mumps.cmake from the YDB repo and is used in a similar manner by CMakeLists.txt to ensure that the shared library is built for both M and UTF-8 modes.