[#5] Fix rare bug where CMake attempts to build _ydbposix.so in UTF-8 mode before the M mode build has finished
This commit fixes a rare bug where CMake attempts to build both versions of _ydbposix.so at the same time without waiting for the other one to finish. We discovered this bug in an internal build where the YDBPosix plugin did not work at all. The fix is to make the M mode version of _ydbposix.so a dependency for the UTF-8 version which ensures that CMake cannot start building the UTF-8 version of _ydbposix.so until after the M mode version has already been built.