[YottaDB/Utils/YDBCMake#5] CMake files refactoring
This commit re-writes the CMakeLists.txt in accordance with the
instructions in YottaDB/Utils/YDBCMake#5 and removes ydbposixso.cmake
as it's not needed anymore. The `ydbcmake` folder is removed as that's
downloaded dynamically. Everything still works as before, so no
instructions change.

`needs_copyright.sh` and `pre-commit` were updated from the Octo sources
as they were erroring out on files that were deleted. This was a problem
we have since fixed.

`git` was added to the pipeline, as it's now needed to download `ydbcmake`.