[YottaDB/DB/YDB#1148] Update .gitlab-ci.yml from Ubuntu 22.04 to 24.04
Small cosmetic changes to `.gitlab-ci.yml` file as well (changing images
to the proper one without an entrypoint, and moving images to each job
rather than applying it globally).

Images changed from `yottadb/yottadb:latest-master` to
`yottadb/yottadb-base:latest-master` as the latter by definition does
not have an entrypoint, and is designed to be used with pipelines for
that very reason.

On Ubuntu 24.04, `libclang-rt-14-dev` is now required for ASAN jobs.